Huge overhaul:
Separated reader and non-reader types (in the future, reader types
should be pub(crate) or even more restricted, once that is stabilized
in Rust. The reading is no longer done sequentially:

The file is read in memory and a cursor starts reading the data. The
dex map file is no longer needed (at least for now). The offset map
has also been removed.

Code items and debug information left.